+++ R/tm_g_bivariate.R
@@ -181,7 +181,7 @@ tm_g_bivariate <- function(label = "Bivariate Plots",
     server = srv_g_bivariate,
     ui = ui_g_bivariate,
     ui_args = args,
-    server_args = data_extract_list,
+    server_args = c(data_extract_list, list(plot_height = plot_height)),
     filters = get_extract_datanames(data_extract_list)
   )
 }
@@ -194,7 +194,7 @@ ui_g_bivariate <- function(id, ...) {
   ns <- NS(id)
   standard_layout(
     output = white_small_well(
-      tags$div(plot_height_output(id = ns("myplot")))
+      tags$div(plot_with_settings_ui(id = ns("myplot"), height = args$plot_height))
     ),
     encoding = div(
       tags$label("Encodings", class = "text-primary"),
@@ -277,7 +277,6 @@ ui_g_bivariate <- function(id, ...) {
           )
         )
       },
-      plot_height_input(id = ns("myplot"), value = args$plot_height),
       panel_group(
         panel_item(
           title = "Plot settings",
@@ -318,7 +317,8 @@ srv_g_bivariate <- function(input,
                             color_settings = FALSE,
                             color,
                             fill,
-                            size) {
+                            size,
+                            plot_height) {
   init_chunks()
   data_extract <- setNames(
     list(x, y),
@@ -362,13 +362,13 @@ srv_g_bivariate <- function(input,
   )
 
   callModule(
-    plot_with_height,
+    plot_with_settings_srv,
     id = "myplot",
-    plot_height = reactive(input$myplot),
-    plot_id = session$ns("plot")
+    plot_r = plot_r,
+    height = plot_height
   )
 
-  output$plot <- renderPlot({
+  plot_r <- reactive({
     chunks_reset()
     chunks_push_data_merge(merged_data())