+++ R/tm_g_response.R
@@ -134,7 +134,7 @@ tm_g_response <- function(label = "Response Plot",
     server = srv_g_response,
     ui = ui_g_response,
     ui_args = args,
-    server_args = data_extract_list,
+    server_args = c(data_extract_list, list(plot_height = plot_height)),
     filters = get_extract_datanames(data_extract_list)
   )
 }
@@ -145,7 +145,7 @@ ui_g_response <- function(id, ...) {
 
   standard_layout(
     output = white_small_well(
-      plot_height_output(id = ns("myplot"))
+      plot_with_settings_ui(id = ns("myplot"), height = args$plot_height)
     ),
     encoding = div(
       tags$label("Encodings", class = "text-primary"),
@@ -181,7 +181,6 @@ ui_g_response <- function(id, ...) {
         selected = ifelse(args$freq, "frequency", "density"),
         justified = TRUE
       ),
-      plot_height_input(id = ns("myplot"), value = args$plot_height),
       panel_group(
         panel_item(
           title = "Plot settings",
@@ -207,7 +206,8 @@ srv_g_response <- function(input,
                            response,
                            x,
                            row_facet,
-                           col_facet) {
+                           col_facet,
+                           plot_height) {
   init_chunks()
   data_extract <- list(response, x, row_facet, col_facet)
   names(data_extract) <- c("response", "x", "row_facet", "col_facet")
@@ -219,12 +219,12 @@ srv_g_response <- function(input,
     input_id = names(data_extract)
   )
 
-  # Insert the plot into a plot_height module from teal.devel
+  # Insert the plot into a plot_with_settings module from teal.devel
   callModule(
-    plot_with_height,
+    plot_with_settings_srv,
     id = "myplot",
-    plot_height = reactive(input$myplot),
-    plot_id = session$ns("plot")
+    plot_r = plot_r,
+    height = plot_height
   )
 
   ## dynamic plot height
@@ -234,7 +234,7 @@ srv_g_response <- function(input,
     plotOutput(session$ns("plot"), height = plot_height)
   })
 
-  output$plot <- renderPlot({
+  plot_r <- reactive({
     chunks_reset()
     chunks_push_data_merge(merged_data())